LayoutTests/http/tests/preload/single_download_preload_runner.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    if (window.internals) {
        internals.setLinkPreloadSupport(true);
        internals.setResourceTimingSupport(true);
    }
    var t = async_test('Makes sure that preloaded resources are not downloaded again when used');

    window.addEventListener("message", t.step_func(function(event) {
        var data = event.data;
        assert_equals(data, 10, "The number of resource downloads");
        if (window.internals) {
            internals.setLinkPreloadSupport(false);
            internals.setResourceTimingSupport(false);
        }
        t.done();
    }));
</script>
<iframe src="resources/single_download_preload.html" onerror="assert_unreached();t.done()"></iframe>
